chapter 4 source code fix
