+++ 04_Building Convolutional Neural Networks/sourceCode/cookbookapp/.gitignore
@@ -7,3 +7,4 @@ cookbookapp.iml
 dependency-reduced-pom.xml
 model.zip
 LocalExecuteExample.csv
+cookbookapp-cnn.iml
