+++ 04_Building Convolutional Neural Networks/sourceCode/cookbookapp/pom.xml
@@ -37,7 +37,7 @@
                     <archive>
                         <manifest>
                             <mainClass>
-                                com.javadeeplearningcookbook.examples.CustomerRetentionPredictionExample
+                                com.javadeeplearningcookbook.examples.AnimalClassifier
                             </mainClass>
                         </manifest>
                     </archive>
@@ -59,7 +59,7 @@
                                         implementation="org.apache.maven.plugins.shade.resource.ApacheLicenseResourceTransformer" />
                                 <transformer
                                         implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                    <mainClass>com.javadeeplearningcookbook.examples.CustomerRetentionPredictionExample</mainClass>
+                                    <mainClass>com.javadeeplearningcookbook.examples.AnimalClassifier</mainClass>
                                 </transformer>
                             </transformers>
                         </configuration>
