+++ 04_Building Convolutional Neural Networks/sourceCode/spring-dl4j/.gitignore
@@ -27,3 +27,15 @@ HELP.md
 
 ### VS Code ###
 .vscode/
+
+.idea
+*.tmp
+target
+cookbook-app.iml
+cookbook-app.iws
+cookbook-app.ipr
+cookbookapp.iml
+cookbookapp-cnn.iml
+dependency-reduced-pom.xml
+model.zip
+LocalExecuteExample.csv
