+++ 04_Building Convolutional Neural Networks/sourceCode/spring-dl4j/pom.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 	<parent>
 		<groupId>org.springframework.boot</groupId>
@@ -29,7 +29,7 @@
 		</dependency>
 		<dependency>
 			<groupId>com.javadeeplearningcookbook.app</groupId>
-			<artifactId>cookbookapp</artifactId>
+			<artifactId>cookbookapp-cnn</artifactId>
 			<version>1.0-SNAPSHOT</version>
 		</dependency>
