+++ 04_Building Convolutional Neural Networks/sourceCode/spring-dl4j/src/main/java/com/springdl4j/springdl4j/SpringDl4jApplication.java
@@ -3,6 +3,7 @@ package com.springdl4j.springdl4j;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
+//Run this only after building project 'cookbookapp-cnn' and jar file is created.
 @SpringBootApplication
 public class SpringDl4jApplication {
