+++ 04_Building Convolutional Neural Networks/sourceCode/spring-dl4j/src/main/java/com/springdl4j/springdl4j/service/CookBookServiceImpl.java
@@ -1,12 +1,13 @@
 package com.springdl4j.springdl4j.service;
 
-import com.javadeeplearningcookbook.api.CustomerRetentionPredictionApi;
+import com.javadeeplearningcookbook.api.ImageClassifierAPI;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import java.io.IOException;
+import java.text.DecimalFormat;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,14 +19,16 @@ public class CookBookServiceImpl implements CookBookService {
         final List<String> results = new ArrayList<>();
         File convFile = File.createTempFile(multipartFile.getOriginalFilename(),null, new File(System.getProperty("user.dir")+"/"));
         multipartFile.transferTo(convFile);
-        INDArray indArray = AnimalClassifierAPI.generateOutput(convFile);
+        INDArray indArray = ImageClassifierAPI.generateOutput(convFile);
+        DecimalFormat df2 = new DecimalFormat("#.####");
         for(int i=0; i<indArray.rows();i++){
-            if(indArray.getDouble(i,0)>indArray.getDouble(i,1)){
-                results.add("Customer "+(i+1)+"-> Happy Customer \n");
+            String result="Image "+String.valueOf(i)+"->>>>>";
+            for(int j=0;j<indArray.columns();j++){
+                result+="\n Category "+j+": "+df2.format(indArray.getDouble(i,j)*100)+"%,   ";
             }
-            else{
-                results.add("Customer "+(i+1)+"-> Unhappy Customer \n");
-            }
+            result+="\n\n";
+            results.add(result);
+
         }
         convFile.deleteOnExit();
